Refactor doc-comment helpers
