+++ src/plugin/src/printer/doc-comment/doc-comment-wrap-width.ts
