+++ src/plugin/src/printer/doc-comment/index.ts
@@ -1,0 +1,4 @@
+export * from "./doc-comment-options.js";
+export * from "./doc-comment-wrap-width.js";
+export * from "./function-docs.js";
+export * from "./synthetic-doc-comment-builder.js";
