+++ src/plugin/src/printer/index.ts
@@ -1,6 +1,6 @@
 export { print, applyAssignmentAlignment } from "./print.js";
 export * as DocBuilders from "./doc-builders.js";
-export * as DocCommentWrapWidth from "./doc-comment-wrap-width.js";
+export * as DocComment from "./doc-comment/index.js";
 export * as LoopSizeHoisting from "./loop-size-hoisting.js";
 export * as SourceText from "./source-text.js";
 export * as StatementSpacingPolicy from "./statement-spacing-policy.js";
